#05: handled errors in backend, dockerized the app, added a database seeding script

+++ todo-list-nuxt-app/package/TaskRepository.ts
@@ -6,6 +6,15 @@ export type todolist = {
     created_date: string
 }
 
+class CustomError extends Error {
+    code: string;
+
+    constructor(message: string, code: string) {
+        super(message);
+        this.code = code;
+    }
+}
+
 export const addTask = async (data: Pick<todolist, 'title'>) => {
     try {
         const result = await sql({
@@ -16,25 +25,51 @@ export const addTask = async (data: Pick<todolist, 'title'>) => {
         return result;
     } catch (error) {
         console.error('Error adding task:', error);
-        throw error;
+        if((error as { code: string }).code === 'ER_NO_SUCH_TABLE') {
+            throw new CustomError('Table not found', 'ER_NO_SUCH_TABLE');
+        }
+        else{
+            throw new CustomError('Internal server error', 'UNHANDLED_ERROR');
+        }
     }
 }
 
 export const deleteTask = async (data: Pick<todolist, 'id'>) => {
     try {
+        const check = await sql({
+            query: 'SELECT * FROM tasks WHERE id = ?',
+            values: [data.id]
+        });
+        if(Array.isArray(check) && check.length === 0){
+            throw new CustomError('Task not found', 'NOT_FOUND');
+        } 
+
         const result = await sql({
             query: 'DELETE FROM tasks WHERE id = ?',
             values: [data.id]
         });
         return result;
     } catch (error) {
-        console.error('Error deleting task:', error);
-        throw error;
+        console.error('Error getting task:', error);
+        if((error as { code: string }).code === 'ER_NO_SUCH_TABLE') {
+            throw new CustomError('Table not found', 'ER_NO_SUCH_TABLE');
+        }
+        else{
+            throw new CustomError('Internal server error', 'UNHANDLED_ERROR');
+        }
     }
 };
 
 export const updateTask = async (data: Pick<todolist, 'id'>) => {
     try {
+        const check = await sql({
+            query: 'SELECT * FROM tasks WHERE id = ?',
+            values: [data.id]
+        });
+        if(Array.isArray(check) && check.length === 0){
+            throw new CustomError('Task not found', 'NOT_FOUND');
+        } 
+
         await sql({
             query: `
                 UPDATE tasks
@@ -46,8 +81,13 @@ export const updateTask = async (data: Pick<todolist, 'id'>) => {
 
         return { message: 'Task updated successfully' };
     } catch (error) {
-        console.error('Error updating task:', error);
-        throw error;
+        console.error('Error getting task:', error);
+        if((error as { code: string }).code === 'ER_NO_SUCH_TABLE') {
+            throw new CustomError('Table not found', 'ER_NO_SUCH_TABLE');
+        }
+        else{
+            throw new CustomError('Internal server error', 'UNHANDLED_ERROR');
+        }
     }
 };
 
@@ -59,7 +99,12 @@ export const getTask = async () => {
     
         return result;
     } catch (error) {
-        console.error('Error fetching tasks:', error);
-        throw error;
+        console.error('Error getting task:', error);
+        if((error as { code: string }).code === 'ER_NO_SUCH_TABLE') {
+            throw new CustomError('Table not found', 'ER_NO_SUCH_TABLE');
+        }
+        else{
+            throw new CustomError('Internal server error', 'UNHANDLED_ERROR');
+        }
     }
 }
